make State lighter, move extra imports
