+++ catalog/app/components/JsonEditor/constants.ts
@@ -1,0 +1,8 @@
+export type JsonValue = $TSFixMe
+
+export const COLUMN_IDS = {
+  KEY: 'key',
+  VALUE: 'value',
+}
+
+export const EMPTY_VALUE = Symbol('empty')
